rename files and reorganize files
