+++ airbyte/_future_cdk/sql_processor.py
@@ -60,7 +60,7 @@ if TYPE_CHECKING:
     )
 
     from airbyte._batch_handles import BatchHandle
-    from airbyte._future_cdk.catalog_managers import CatalogProvider
+    from airbyte._future_cdk.catalog_providers import CatalogProvider
     from airbyte._future_cdk.state_writers import StateWriterBase
     from airbyte._processors.file.base import FileWriterBase
     from airbyte.secrets.base import SecretString
